Create WordPress_RSS_Grab
This code should be added to the functions.php file within the theme or child_theme folder of a WordPress site.  The purpose of this function is to grab the images from an rss feed associated with (in this case) a pinterest account. Note: amend line 7, replace myPinterestAccount with the name of an actual pinterest account. 

When ready, copy the shortcode [pinterest_feed] and paste it into the page, post or widget of your choice.